AI_Agent/requirements.txt
# =============================================================================
# Multi-Agent System for Wind Pressure Cp Time Series Modeling
# Requirements
# =============================================================================

# --- LLM & Agent Framework ---
langchain>=0.3.0
langgraph>=0.2.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

# --- Literature Search APIs ---
arxiv>=2.1.0
semanticscholar>=0.8.0
google-search-results>=2.4.2   # SerpAPI for Google Scholar

# --- PDF Parsing ---
PyMuPDF>=1.24.0                # fitz for PDF text extraction

# --- Deep Learning ---
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.40.0
einops>=0.8.0

# --- ML & Statistics ---
scikit-learn>=1.4.0
statsmodels>=0.14.0
xgboost>=2.0.0
lightgbm>=4.3.0
optuna>=3.6.0

# --- Experiment Tracking ---
mlflow>=2.14.0

# --- Data Processing ---
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.12.0

# --- Visualization ---
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# --- Configuration & Environment ---
pyyaml>=6.0.1
python-dotenv>=1.0.1

# --- Utilities ---
joblib>=1.3.0
tqdm>=4.66.0
rich>=13.7.0                   # Console output formatting
psutil>=5.9.0
requests>=2.31.0
aiohttp>=3.9.0                 # Async HTTP for parallel searches
tenacity>=8.2.0                # Retry logic for API calls
